requirements.txt
django
djangorestframework
wordcloud
tqdm
janome